Move StartetPlaylist to appropriate folder
