+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/HostGameHandler.kt
@@ -1,11 +1,13 @@
 package com.github.fribourgsdp.radio.game.handler
 
 import android.util.Log
-import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.database.FirestoreDatabase
 import com.github.fribourgsdp.radio.game.Game
 import com.github.fribourgsdp.radio.game.GameView
+import com.github.fribourgsdp.radio.util.getPlayerDoneMap
+import com.github.fribourgsdp.radio.util.getPlayerFoundMap
+import com.github.fribourgsdp.radio.util.getScoresOfRound
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.FieldValue
 import java.lang.IllegalStateException
