+++ app/src/main/java/com/github/fribourgsdp/radio/game/prep/LobbyActivity.kt
@@ -23,6 +23,8 @@ import com.github.fribourgsdp.radio.external.musixmatch.MusixmatchLyricsGetter
 import com.github.fribourgsdp.radio.game.Game
 import com.github.fribourgsdp.radio.game.GameActivity
 import com.github.fribourgsdp.radio.game.view.QuitGameOrLobbyDialog
+import com.github.fribourgsdp.radio.util.getPermissions
+import com.github.fribourgsdp.radio.util.getPlayers
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
